# [Utilities] improve performance of canonical